correct format + TU. some prac.
